chore: add resuable FeeDetails component
