+++ apps/tangle-dapp/app/bridge/BridgeContainer.tsx
@@ -1,12 +1,9 @@
 'use client';
 
 import Button from '@webb-tools/webb-ui-components/components/buttons/Button';
-import FeeDetails from '@webb-tools/webb-ui-components/components/FeeDetails';
-import type { FeeItem } from '@webb-tools/webb-ui-components/components/FeeDetails/types';
 import InfoIconWithTooltip from '@webb-tools/webb-ui-components/components/IconWithTooltip/InfoIconWithTooltip';
 import { Typography } from '@webb-tools/webb-ui-components/typography/Typography';
-import Decimal from 'decimal.js';
-import { FC, useMemo, useState } from 'react';
+import { FC, useState } from 'react';
 import { twMerge } from 'tailwind-merge';
 
 import AddressInput, {
@@ -17,11 +14,8 @@ import { isEVMChain } from '../../utils/bridge';
 import AmountAndTokenInput from './AmountAndTokenInput';
 import BridgeConfirmationModal from './BridgeConfirmationModal';
 import ChainSelectors from './ChainSelectors';
+import FeeDetails from './FeeDetails';
 import useActionButton from './hooks/useActionButton';
-import useBridgeFee from './hooks/useBridgeFee';
-import useEstimatedGasFee from './hooks/useEstimatedGasFee';
-import useSelectedToken from './hooks/useSelectedToken';
-import useTypedChainId from './hooks/useTypedChainId';
 
 interface BridgeContainerProps {
   className?: string;
@@ -29,7 +23,6 @@ interface BridgeContainerProps {
 
 const BridgeContainer: FC<BridgeContainerProps> = ({ className }) => {
   const {
-    selectedSourceChain,
     selectedDestinationChain,
     destinationAddress,
     setDestinationAddress,
@@ -38,42 +31,6 @@ const BridgeContainer: FC<BridgeContainerProps> = ({ className }) => {
   const [isConfirmModalOpen, setIsConfirmModalOpen] = useState(false);
   const { buttonAction, buttonText, isLoading, isDisabled, errorMessage } =
     useActionButton(() => setIsConfirmModalOpen(true));
-  const { fee: bridgeFee, isLoading: isLoadingBridgeFee } = useBridgeFee();
-  const { fee: estimatedGasFee, isLoading: isLoadingEstimatedGasFee } =
-    useEstimatedGasFee();
-  const selectedToken = useSelectedToken();
-  const { destinationTypedChainId } = useTypedChainId();
-
-  const destChainTransactionFee = useMemo(
-    () =>
-      selectedToken.destChainTransactionFee[destinationTypedChainId] ?? null,
-    [destinationTypedChainId, selectedToken.destChainTransactionFee],
-  );
-
-  const totalFeeCmp = useMemo(() => {
-    if (bridgeFee === null || estimatedGasFee === null) return null;
-
-    const allTokenFee = bridgeFee.add(
-      destChainTransactionFee ?? new Decimal(0),
-    );
-
-    if (
-      selectedToken.symbol.toLowerCase() ===
-      selectedSourceChain.nativeCurrency.symbol.toLowerCase()
-    ) {
-      const totalFee = allTokenFee.add(estimatedGasFee);
-      if (totalFee.isZero()) return null;
-      return `${totalFee.toString()} ${selectedToken.symbol}`;
-    }
-
-    return `${allTokenFee.toString()} ${selectedToken.symbol} + ${estimatedGasFee.toString()} ${selectedSourceChain.nativeCurrency.symbol}`;
-  }, [
-    bridgeFee,
-    destChainTransactionFee,
-    selectedToken.symbol,
-    estimatedGasFee,
-    selectedSourceChain.nativeCurrency.symbol,
-  ]);
 
   return (
     <>
@@ -108,55 +65,7 @@ const BridgeContainer: FC<BridgeContainerProps> = ({ className }) => {
               }
             />
 
-            <FeeDetails
-              title="Total Fees"
-              totalFeeCmp={totalFeeCmp}
-              isTotalLoading={isLoadingBridgeFee || isLoadingEstimatedGasFee}
-              items={
-                [
-                  bridgeFee !== null
-                    ? {
-                        name: 'Bridge Fee',
-                        value: (
-                          <FeeValueCmp
-                            fee={bridgeFee}
-                            symbol={selectedToken.symbol}
-                          />
-                        ),
-                        isLoading: isLoadingBridgeFee,
-                        info: 'This transaction will charge a bridge fee to cover the destination chain’s gas fee.',
-                      }
-                    : undefined,
-                  destChainTransactionFee !== null
-                    ? {
-                        name: 'Bridge Fee',
-                        value: (
-                          <FeeValueCmp
-                            fee={destChainTransactionFee}
-                            symbol={selectedToken.symbol}
-                          />
-                        ),
-                        info: 'This fee is used to pay the XCM fee of the destination chain.',
-                      }
-                    : undefined,
-                  estimatedGasFee !== null
-                    ? {
-                        name: 'Estimated Gas Fee',
-                        value: (
-                          <FeeValueCmp
-                            fee={estimatedGasFee}
-                            symbol={selectedSourceChain.nativeCurrency.symbol}
-                          />
-                        ),
-                        isLoading: isLoadingEstimatedGasFee,
-                      }
-                    : undefined,
-                ].filter((item) => Boolean(item)) as Array<FeeItem>
-              }
-              className="!bg-mono-20 dark:!bg-mono-160"
-              titleClassName="!text-mono-100 dark:!text-mono-80"
-              itemTitleClassName="!text-mono-100 dark:!text-mono-80"
-            />
+            <FeeDetails />
           </div>
           <div className="flex flex-col items-end gap-2">
             {errorMessage && (
@@ -198,12 +107,3 @@ const BridgeContainer: FC<BridgeContainerProps> = ({ className }) => {
 };
 
 export default BridgeContainer;
-
-const FeeValueCmp: FC<{ fee: Decimal; symbol: string }> = ({ fee, symbol }) => {
-  return (
-    <Typography
-      variant="body1"
-      className="!text-mono-120 dark:!text-mono-100"
-    >{`${fee.toString()} ${symbol}`}</Typography>
-  );
-};
